requirements.txt
python-dotenv

pynacl		~=1.4.0
discord.py	~=1.7.3
youtube_dl	~=2021.6.6
